+++ foxy_farmer/farmer/chia_farmer.py
@@ -1,8 +1,11 @@
 from abc import ABC
-from typing import List
+from asyncio import gather
+from pathlib import Path
+from typing import List, Awaitable
 
 from foxy_farmer.environment.chia_environment import ChiaEnvironment
 from foxy_farmer.farmer.farmer import Farmer
+from foxy_farmer.monitoring.monitor_farmer import monitor_farmer
 
 
 class ChiaFarmer(Farmer, ABC):
@@ -15,13 +18,17 @@ class ChiaFarmer(Farmer, ABC):
         return services
 
     _environment: ChiaEnvironment
+    _root_path: Path
 
     async def run(self) -> None:
         await self._environment.init()
         try:
             await self._environment.start_daemon()
             await self._environment.start_services(self._services_to_run)
-            await self._stop_event.wait()
+            futures: List[Awaitable] = [self._stop_event.wait()]
+            if self._farmer_config.get("monitor_farmer_connections") is True:
+                futures.append(monitor_farmer(root_path=self._root_path, until=self._stop_event))
+            await gather(*futures)
         finally:
             await self._environment.stop_services(self._services_to_run)
             await self._environment.stop_daemon()