+++ foxy_farmer/farmer/gigahorse_farmer.py
@@ -35,6 +35,7 @@ class GigahorseFarmer(ChiaFarmer):
             allow_connecting_to_existing_daemon=False,
             farmer_config=farmer_config,
         )
+        self._root_path = root_path
         self._syslog_server = SyslogServer(logging_config=config["logging"])
 
     async def run(self) -> None: