+++ foxy_farmer/farmer/split_chia_farmer.py
@@ -1,9 +1,12 @@
 from abc import ABC
-from typing import List
+from asyncio import gather
+from pathlib import Path
+from typing import List, Awaitable
 
 from foxy_farmer.environment.chia_environment import ChiaEnvironment
 from foxy_farmer.environment.embedded_chia_environment import EmbeddedChiaEnvironment
 from foxy_farmer.farmer.farmer import Farmer
+from foxy_farmer.monitoring.monitor_farmer import monitor_farmer
 
 
 class SplitChiaFarmer(Farmer, ABC):
@@ -21,6 +24,7 @@ class SplitChiaFarmer(Farmer, ABC):
 
     _environment: ChiaEnvironment
     _embedded_environment: EmbeddedChiaEnvironment
+    _root_path: Path
 
     async def run(self) -> None:
         daemon_environment = self._daemon_environment
@@ -30,7 +34,10 @@ class SplitChiaFarmer(Farmer, ABC):
             await daemon_environment.start_daemon()
             await self._embedded_environment.start_services(self._services_to_run_on_embedded_env)
             await daemon_environment.start_services(self._services_to_run_on_env)
-            await self._stop_event.wait()
+            futures: List[Awaitable] = [self._stop_event.wait()]
+            if self._farmer_config.get("monitor_farmer_connections") is True:
+                futures.append(monitor_farmer(root_path=self._root_path, until=self._stop_event))
+            await gather(*futures)
         finally:
             await daemon_environment.stop_services(self._services_to_run_on_env)
             await self._embedded_environment.stop_services(self._services_to_run_on_embedded_env)